🐛 Fix dynamic functions in `needextract` need
fixes #1267
